+++ tier2/riff.cpp
@@ -128,7 +128,7 @@ InFileRIFF::~InFileRIFF( void )
 int InFileRIFF::ReadInt( void )
 {
 	int tmp = 0;
-	m_io.read( &tmp, sizeof(intp), m_file );
+	m_io.read( &tmp, sizeof(int), m_file );
 	tmp = LittleLong( tmp );
 
 	return tmp;
